make `sml/replacer' symlink-aware
If I have a directory ~/long/path/to/my/project and then a shortcut
symlink ~/myproject which points to it, and I want sml to replace
pathname prefixes with :MYPRJ:, previously I would have to add two
entries to `sml/replacer-regexp-list': one for the long pathname, and
one for the symlink.

This commit makes sml smart enough to only require one entry for the
long pathname, and then automatically spot that when I've loaded a
project file ~/myproject/foo.py, the absolute (expanded) path matches an
entry in `sml/replacer-regexp-list'.