Describe a workaround for a safe-themes problem.
(sml/apply-theme ‘automatic) can attempt to apply a theme that hasn’t been used yet, and so hasn’t been marked safe. Worse, it can do this when launching emacs in —daemon mode, leading to severe debugging pain.

Add a description of the problem and its workaround to the README.md.