visibility for sml/time
Level 0 background is black.  sml/time font pulls from l0 for background, but sml/modes for foreground.  sml/modes has a foreground of black.  This uses the settings from sml/col-number, to keep the symmetry of the edges the same.  My powerline shows the buffer size on the far left, but I couldn't find that setting in the file.  I chose the one that is applies to sml/modes, it looked the same.