Prevent irrelevant warning about sml/hidden-modes being obsolete
When reloading smart-mode-line.el, then `sml-hidden-modes' is bound
because that file defines it as an obsolete variable after checking
whether the user set it somehow.

Previously that resulted in a warning about it being obsolete, but
that is not appropriate if the user did not actually set that variable.

Fixes #213.